Grid: Rename nemo_grid.py -> grid.py
