+++ parcels/__init__.py
@@ -1,3 +1,3 @@
-from parcels.nemo_grid import *
+from parcels.grid import *
 from parcels.particle import *
 from parcels.field import *
